Refactor: Migrate media item IDs from String to Long

+++ shared/compose/usecase/src/commonMain/kotlin/com/andannn/melodify/shared/compose/usecase/CustomTabContentUseCase.kt
@@ -58,7 +58,7 @@ fun Tab.contentFlow(
 
         is Tab.PlayListDetail -> {
             repository.getItemsOfPlayListFlow(
-                playListId.toLong(),
+                playListId,
                 sorts.filterIsInstance<SortOption.PlayListOption>(),
                 whereGroups,
             )
@@ -112,7 +112,7 @@ fun Tab.contentPagingDataFlow(
 
         is Tab.PlayListDetail -> {
             repository.getItemsPagingFlowOfPlayList(
-                playListId.toLong(),
+                playListId,
                 sorts.filterIsInstance<SortOption.PlayListOption>(),
                 whereGroups,
             )
